refactor(ir): move graph building fns to ConstraintBuilder
